Refactor: GeometryManager
