+++ src/sceneManager.js
@@ -4,12 +4,12 @@ import CameraRig from "./cameraRig.js"
 import MapControlsFactory from './mapControlsFactory.js'
 import RendererFactory from './rendererFactory.js'
 import eventBus from './utils/eventBus.js';
-
+import { genomicService } from './main.js';
 let previousNodeName = undefined
 
 class SceneManager {
 
-    constructor(container, backgroundColor, frustumSize, raycastService, dataService, sequenceService, genomicService) {
+    constructor(container, backgroundColor, frustumSize, raycastService, dataService, sequenceService) {
         this.container = container
         this.scene = new THREE.Scene()
         this.scene.background = backgroundColor
@@ -17,7 +17,7 @@ class SceneManager {
 
         this.dataService = dataService
         this.sequenceService = sequenceService
-        this.genomicService = genomicService
+       
         // Initialize renderer
         this.renderer = RendererFactory.create(container)
 
@@ -50,7 +50,6 @@ class SceneManager {
     }
 
     handleIntersection(intersections) {
-
         if (undefined === intersections || 0 === intersections.length) {
             this.clearIntersection()
             return
@@ -59,24 +58,23 @@ class SceneManager {
         // Sort by distance to get the closest intersection
         intersections.sort((a, b) => a.distance - b.distance);
 
-		const { faceIndex, pointOnLine, object } = intersections[0];
+        const { faceIndex, pointOnLine, object } = intersections[0];
 
         this.renderer.domElement.style.cursor = 'none';
 
-        const { t, nodeName, nodeLine } = this.raycastService.handleIntersection(this.dataService, object, pointOnLine, faceIndex);
+        const { t, nodeName, nodeLine } = this.raycastService.handleIntersection(this.dataService.geometryManager, object, pointOnLine, faceIndex);
 
         eventBus.publish('lineIntersection', { t, nodeName, nodeLine })
-
-	}
+    }
 
     clearIntersection() {
-		this.raycastService.clearIntersection()
+        this.raycastService.clearIntersection()
         this.renderer.domElement.style.cursor = '';
-	}
+    }
 
     animate() {
         if (true === this.raycastService.isEnabled) {
-            const intersections = this.raycastService.intersectObject(this.cameraRig.camera, this.dataService.linesGroup)
+            const intersections = this.raycastService.intersectObjects(this.cameraRig.camera, this.dataService.geometryManager.linesGroup.children)
             this.handleIntersection(intersections)
         }
 
@@ -94,7 +92,6 @@ class SceneManager {
     }
 
     updateViewToFitScene() {
-
         // Create a bounding box that encompasses all objects in the scene
         const bbox = new THREE.Box3()
         this.scene.traverse((object) => {
@@ -174,7 +171,7 @@ class SceneManager {
 
         this.dataService.dispose()
 
-        this.dataService.ingestData(json, this.genomicService)
+        this.dataService.ingestData(json, genomicService)
 
         this.dataService.addToScene(this.scene)
